+++ platformio/include/LVGLFunctions.hpp
@@ -129,7 +129,7 @@ void getSampleFunc(lv_task_t *task)
     }
     if (config.currentSampleNumber != 0 && config.currentSampleNumber < config.countOfSamples)
     {
-        std::map<std::string, int16_t> tmpData = pmsSensor->returnData();
+        std::map<std::string, int32_t> tmpData = pmsSensor->returnData();
         pmsSensor->dumpSamples();
         for (uint8_t i = 0; i < 15; i++)
         {
@@ -144,7 +144,7 @@ void getSampleFunc(lv_task_t *task)
         lv_task_set_period(getSample, config.measurePeriod);
         if (pmsSensor->readData())
         {
-            std::map<std::string, int16_t> tmpData = pmsSensor->returnData();
+            std::map<std::string, int32_t> tmpData = pmsSensor->returnData();
             pmsSensor->dumpSamples();
             data = tmpData;
             config.currentSampleNumber++;
