+++ platformio/include/MySD.hpp
@@ -23,7 +23,7 @@ public:
     bool begin();
     void end();
     bool start(SQLiteDb *object, Stream *debugger);
-    void save(std::map<std::string, int16_t> data, int temperature, int humidity, String timestamp, SQLiteDb *object, Stream *debugger);
+    void save(std::map<std::string, int32_t> data, int temperature, int humidity, String timestamp, SQLiteDb *object, Stream *debugger);
     void select(SQLiteDb *object, Stream *debugger, String datetime, JsonArray *array);
     void getLastRecord(SQLiteDb *object, Stream *debugger, JsonArray *array);
     void saveConfig(Config config, std::string filePath);
